apps/ai-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pillow==11.0.0
opencv-python-headless==4.10.0.84
numpy==2.1.0
pydantic==2.10.0
httpx==0.28.0
